services/streaming/requirements.txt
requests==2.28.2
python-dotenv==1.0.0
kafka-python==2.0.2
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
river==0.14.0
Flask==2.2.3
Werkzeug==2.2.3
